requirements.txt
Flask==3.1.0
google_api_python_client==2.151.0
google_auth_oauthlib==1.2.1
openai==1.55.1
pikepdf==9.4.2
protobuf==5.28.3
pymongo==4.10.1
python-dotenv==1.0.1
reportlab==4.2.5
Requests==2.32.3
spacy==3.8.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
